add-custom-alerts-and-fix-error-create-user

+++ client/src/components/ShoppingCart/ShoppingCart.jsx
@@ -3,6 +3,7 @@ import { BsFillTrash3Fill } from "react-icons/bs";
 import { useState } from "react";
 import axios from "axios";
 import { useSelector } from "react-redux";
+import { alertAcept, alertConfirm } from "../SweetAlert/SweetAlert";
 
 const { VITE_SERVER_URL } = import.meta.env;
 
@@ -43,23 +44,45 @@ export const ShoppingCart = () => {
     localStorage.setItem("cartProducts", JSON.stringify(updatedProducts));
   };
 
-  const handleRemoveItem = (productId) => {
-    const updatedProducts = paymentData.products.filter(
-      (product) => product.id !== productId
-    );
-    setPaymentData((prevData) => ({
-      ...prevData,
-      products: updatedProducts,
-    }));
-    localStorage.setItem("cartProducts", JSON.stringify(updatedProducts));
+  const handleRemoveItem = async (productId) => {
+    try {
+      const alert = await alertConfirm(
+        "warning",
+        "Delete product!",
+        "Are you sure you want to remove this product?"
+      );
+      if (alert) {
+        const updatedProducts = paymentData.products.filter(
+          (product) => product.id !== productId
+        );
+        setPaymentData((prevData) => ({
+          ...prevData,
+          products: updatedProducts,
+        }));
+        localStorage.setItem("cartProducts", JSON.stringify(updatedProducts));
+      }
+    } catch (error) {
+      console.error(error.message);
+    }
   };
 
-  const handleRemoveAll = () => {
-    setPaymentData((prevData) => ({
-      ...prevData,
-      products: [],
-    }));
-    localStorage.removeItem("cartProducts");
+  const handleRemoveAll = async () => {
+    try {
+      const alert = await alertConfirm(
+        "warning",
+        "Delete all product!",
+        "Are you sure you want to remove all products?"
+      );
+      if (alert) {
+        setPaymentData((prevData) => ({
+          ...prevData,
+          products: [],
+        }));
+        localStorage.removeItem("cartProducts");
+      }
+    } catch (error) {
+      console.error(error.message);
+    }
   };
 
   return (
@@ -103,15 +126,12 @@ export const ShoppingCart = () => {
 
         <div className={style.totalPrice}>
           Sub Total : $
-          {
-            (paymentData.products
-              .reduce(
-                (total, product) =>
-                  total + product.unit_price * product.quantity,
-                0
-              )
-              .toFixed(2))
-          }
+          {paymentData.products
+            .reduce(
+              (total, product) => total + product.unit_price * product.quantity,
+              0
+            )
+            .toFixed(2)}
           <button disabled={!authData?.id} onClick={() => shopping()}>
             CheckOut
           </button>
